Struct krpsim::format::operate::process::Process

pub struct Process {
    pub name: String,
    pub cycle: usize,
    pub input: Inventory,
    pub output: Inventory,
    neutral: Option<Ressource>,
    pub heuristic: HashMap<String, f64>,
}

Fields
name: String cycle: usize input: Inventory output: Inventory neutral: Option<Ressource> heuristic: HashMap<String, f64>
Methods

impl Process

fn new(name: String,
cycle: usize,
input: Inventory,
output: Inventory,
hash: HashMap<String, f64>)
-> Self

The new constructor function returns the Process.

fn from_integer(name: String,
cycle: usize,
input: Inventory,
output: Inventory)
-> Self

fn from_line(name: String,
need: &str,
result_and_nb_cycle: &str)
-> Result<Process>

The from_line constructor function returns a parsed process.

fn get_name(&self) -> &str

The get_name accessor function returns the name of process.

fn get_cycle(&self) -> usize

The get_cycle accessor function returns the number of cycle required by the process.

fn buy_with(&self, with: &mut Inventory) -> Result<()>

The buy_with function substrates with argument with input and additions the output to with argument.

fn get_h_value(&self, s: &String) -> f64

fn get_distance(need: &Ressource, owned: &Vec<Ressource>) -> usize

fn get_producing_process(obj: &Ressource,
process: &Vec<&Process>,
used_process: Vec<Process>)
-> Vec<Process>

fn get_ressource_number(vect: &Vec<Process>, obj: &Ressource) -> f64

fn time_cmp(ori: &Result<(Vec<Process>, usize), ()>,
new: (&Vec<Process>, usize),
obj: &Ressource)
-> bool

fn needed_process(&self,
process: &Vec<&Process>,
ressources: &mut Inventory,
real_ressources: &mut Inventory,
already_used: Vec<Process>,
delay: usize)
-> Result<(Option<Vec<Process>>, usize), ()>

fn distance_overall(&self, owned: &Vec<Ressource>) -> usize