Struct krpsim::format::operate::running::Running

pub struct Running(HashMap<String, Process>);

Methods

impl Running

fn new(process: Vec<Process>) -> Self

The new constructor function returns the list of process.

fn is_empty(&self) -> bool

The is_empty interface function returns true if the map contains not elements.

fn len(&self) -> usize

The len interface function returns the number of elements in the map.

fn iter(&self) -> Iter<String, Process>

The iter interface function returns a iterator.

fn push(&mut self, key: String, val: Process) -> Option<Process>

The push interface function inserts a new item to the inventory.

fn get(&self, key: &String) -> Option<&Process>

The get_process accesor function returns a process from map according to a key name.

fn buy_with(&self, commands: &Trace, inventory: &mut Inventory) -> Result<()>

The buy_with function buys one command with an inventary.

fn can_cycle(&self, checks: &Trace) -> Result<usize>

The can_cycle checks if the number and name of cycle is right between two process.

fn get_process(&self) -> Vec<&Process>

The get_process function returns a accessor on the list of process.

fn get_cloned_process(&self) -> Vec<Process>