Struct krpsim::format::optimize::Optimize

pub struct Optimize {
    pub stock: Vec<String>,
    pub time: bool,
}

The Optimize structure is a list of keywords/stock-names who have the priority queue.

Fields
stock: Vec<String> time: bool
Methods

impl Optimize

fn new(stock: Vec<String>, time: bool) -> Self

The new constructor function returns the optimization's items.

fn from_line(line: String) -> Self

The from_line constructor function returns the optimization's item for a parsed line.

fn len(&self) -> usize

The len interface function returns the number of elements in the list.

fn is_empty(&self) -> bool

The is_empty interface function returns true if the vector Optimize contains no elements.