Struct krpsim::format::stock::inventory::Inventory

pub struct Inventory(HashMap<String, Ressource>);

Methods

impl Inventory

fn new(ressources: Vec<Ressource>) -> Self

The new constructor function returns the list of ressources.

fn from_result(ressources: Vec<Result<Ressource>>) -> Option<Self>

The from_result multi constructor function returns a list of Item for a list of Result.

fn print_final(&self)

fn from_line(ressources: &str) -> Option<Self>

The from_line multi constructor function returns a list of Item for a need or result argument from Process.

fn len(&self) -> usize

The len interface function returns the number of elements in the map.

fn iter(&self) -> Iter<String, Ressource>

The iter interface function returns a iterator.

fn push(&mut self, key: String, val: Ressource) -> Option<Ressource>

The push interface function inserts a new item to the inventory.

fn is_empty(&self) -> bool

The is_empty interface function returns true if the map contains not elements.

fn is_zero(&self) -> bool

The is_zero interface function returns true if the map contains only nul ressources

fn any(&self, key: &str) -> bool

The any interface function checks if the map contains the key.

fn any_from_ressource(&self, val: &Ressource) -> bool

The any interface function checks if the map contains the key from a ressource.

fn get(&self, key: &str) -> Option<&Ressource>

The get accessor function returns a item.

fn get_from_ressource(&self, val: &Ressource) -> Option<&Ressource>

The get_from_ressource accessor function returns a item.

fn get_mut(&mut self, key: &str) -> Option<&mut Ressource>

The get_mut mutator function returns a item.

fn get_mut_from_ressource(&mut self, val: &Ressource) -> Option<&mut Ressource>

The get_mut mutator function returns a item.

fn add(&mut self, val: &Ressource) -> Option<usize>

The add interface function additiones a item with another item.

fn add_from_inventory(&mut self, vals: &Inventory) -> bool

The add interface function additiones a list of item to our self-inventory.

fn sub(&mut self, val: &Ressource) -> Option<usize>

The sub interface function substractes a item with another item.

fn sub_from_inventory(&mut self, vals: &Inventory) -> bool

The sub interface function substractes a list of item to our self-inventory.

fn order(&self, with: &mut Inventory) -> Result<()>

The order takes the payment of command.

fn can_order(&self, with: &Inventory) -> Result<()>

The can_order checks if the order is possible.

fn get_ressource(&self) -> Vec<&Ressource>

The get_ressource function returns a accessor on the list of ressource.

fn get_neutral(&self, output: &Inventory) -> Option<Ressource>

The get_neutral function returns return a neutral component if the output ressource exist.

fn to_vec(&self) -> Vec<Ressource>

Deprecated

The to_vec function returns a cloned list of ressource.