Module krpsim::parser::config

The module config describes a configuration of industrial process.

Reexports

extern crate std;

use itertools::Itertools;

use itertools::FoldWhile::{Continue, Done};

use format::stock::ressource::Ressource;

use format::stock::inventory::Inventory;

use format::operate::process::Process;

use format::operate::running::Running;

use format::optimize::Optimize;

use std::io::prelude::*;

Structs

Configuration
The Configuration struct contains the Ressource, Process and Optimize.

Constants

ERR_DOUBLE

ERR_READ

ERR_SPLITN