Struct krpsim::parser::config::Configuration

pub struct Configuration {
    pub ressources: Inventory,
    pub running: Running,
    pub optimize: Optimize,
}

The Configuration struct contains the Ressource, Process and Optimize.

Fields
ressources: Inventory running: Running optimize: Optimize
Methods

impl Configuration

fn new(file: &str) -> Result<Self>

The new constructor function returns a parsed struct of Ressource, Process and Optimize.