Struct krpsim::parser::trace::Trace

pub struct Trace(Vec<(String, usize)>);

Methods

impl Trace

fn new(file: &str) -> Result<Self>

The new constructor function returns a list of order by cycle.

fn from_vec(list: Vec<(String, usize)>) -> Self

The from_vec constructor function returns a list of order by cycle according to a vector.

fn iter(&self) -> Iter<(String, usize)>

The iter interface function returns a iterator.

fn push(&mut self, product: String, quantity: usize)

The push interface function inserts a new order to the list.